+++ app/src/main/java/com/example/skite/ui/screens/sessionScreen/SessionDetailScreen.kt
@@ -155,6 +155,13 @@ fun SessionDetailScreenContent(
                             }
                         }
                     }
+                    else -> {
+                        //TODO set up cancel screen
+                        ErrorContent(
+                            message = stringResource(R.string.TBD),
+                            onRetry = onRetry
+                        )
+                    }
                 }
             }
         }