+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [Unreleased]
 
-_Description: Écrire le résumé ici..._
+_Added Cancel session status_
 
 <!--
 BUMP_TYPE :
@@ -10,12 +10,12 @@ BUMP_TYPE :
 2 = Minor (0.X.0)
 3 = Patch (0.0.X)
 -->
-Bump: [Numéro]
+Bump: 3
 
 ### Features
 
 ### Patches
-
+- Session state added cancel
 ### Bug Fixes
 
 ### Deployment & Configuration